+++ common/src/main/java/net/forcemaster_rpg/item/armor/BillporonArmor.java
@@ -2,7 +2,7 @@ package net.forcemaster_rpg.item.armor;
 
 import net.minecraft.item.ArmorMaterial;
 import net.minecraft.registry.entry.RegistryEntry;
-import net.spell_engine.api.item.armor.Armor;
+import net.spell_engine.rpg_series.item.Armor;
 
 public class BillporonArmor extends Armor.CustomItem {
     public BillporonArmor(RegistryEntry<ArmorMaterial> material, Type slot, Settings settings) {