+++ common/src/main/java/net/forcemaster_rpg/spell/ForcemasterSpells.java
@@ -1,6 +1,7 @@
 package net.forcemaster_rpg.spell;
 
 import net.forcemaster_rpg.effect.ForcemasterEffects;
+import net.forcemaster_rpg.sounds.ModSounds;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Identifier;
@@ -9,11 +10,13 @@ import net.spell_engine.api.datagen.SpellBuilder;
 import net.spell_engine.api.spell.ExternalSpellSchools;
 import net.spell_engine.api.spell.Spell;
 import net.spell_engine.api.spell.fx.ParticleBatch;
+import net.spell_engine.api.spell.fx.PlayerAnimation;
 import net.spell_engine.api.spell.fx.Sound;
 import net.spell_engine.api.util.TriState;
 import net.spell_engine.client.gui.SpellTooltip;
 import net.spell_engine.client.util.Color;
 import net.spell_engine.fx.SpellEngineParticles;
+import net.spell_engine.fx.SpellEngineSounds;
 import net.spell_engine.internals.target.SpellTarget;
 import net.spell_power.api.SpellSchools;
 import org.jetbrains.annotations.Nullable;
@@ -26,8 +29,22 @@ import java.util.List;
 import static net.forcemaster_rpg.ForcemasterClassMod.MOD_ID;
 
 public class ForcemasterSpells {
+    public enum Book { FORCEMASTER }
     public record Entry(Identifier id, Spell spell, String title, String description,
-                        @Nullable SpellTooltip.DescriptionMutator mutator) {
+                        @Nullable SpellTooltip.DescriptionMutator mutator,
+                        @Nullable Book book) {
+        public Entry(Identifier id, Spell spell, String title, String description) {
+            this(id, spell, title, description, null, null);
+        }
+        public Entry(Identifier id, Spell spell, String title, String description, @Nullable SpellTooltip.DescriptionMutator mutator) {
+            this(id, spell, title, description, mutator, null);
+        }
+        public Entry mutator(SpellTooltip.DescriptionMutator mutator) {
+            return new Entry(id, spell, title, description, mutator, book);
+        }
+        public Entry book(Book book) {
+            return new Entry(id, spell, title, description, mutator, book);
+        }
     }
 
     public static final List<Entry> entries = new ArrayList<>();
@@ -100,50 +117,65 @@ public class ForcemasterSpells {
         return triggers;
     }
 
-
-    ///PASSIVES
-    public static final Entry knuckle_knockup = add(knuckle_knockup());
-    private static Entry knuckle_knockup() {
-        var id = Identifier.of(MOD_ID, "knuckle_knockup");
-        var title = "Knuckle Knock Up";
-        var description = "Hitting enemies has a {trigger_chance} chance to knock up the target.";
-        var spell = SpellBuilder.createSpellPassive();
-        spell.tooltip = new Spell.Tooltip();
-        spell.tooltip.show_header = false;
-        spell.tooltip.name = new Spell.Tooltip.LineOptions(false, false);
-        spell.tooltip.description.color = Formatting.DARK_GREEN.asString();
-        spell.tooltip.description.show_in_compact = false;
+    /// WEAPON SKILL
+    public static final Entry burstcrack = add(burstcrack());
+    private static Entry burstcrack() {
+        var id = Identifier.of(MOD_ID, "burstcrack");
+        var title = "Burstcrack";
+        var description = "Unleashes a shockwave around you, knocking up enemies and dealing physical {damage_1} & {damage_2} arcane-damage.";
+        var spell = SpellBuilder.createWeaponSpell();
         spell.school = ExternalSpellSchools.PHYSICAL_MELEE;
-        spell.range = 0;
+        spell.range = 5.5F;
 
-        spell.target.type = Spell.Target.Type.FROM_TRIGGER;
+        spell.active.cast.movement_speed = 0.25F;
+        spell.active.cast.duration = 0.5F;
+        spell.active.cast.animation = PlayerAnimation.of("forcemaster_rpg:burstcrack_cast");
 
-        var trigger = SpellBuilder.Triggers.meleeAttackImpact();
-        trigger.chance = 0.15F;
-        spell.passive.triggers = List.of(trigger);
+        spell.release.animation = PlayerAnimation.of("forcemaster_rpg:burstcrack_release");
+        spell.release.sound = Sound.withVolume(Identifier.of("entity.generic.explode"), 0.4F);
+        spell.release.particles = new ParticleBatch[]{
+                new ParticleBatch(SpellEngineParticles.smoke_medium.id().toString(),
+                        ParticleBatch.Shape.CIRCLE, ParticleBatch.Origin.CENTER,
+                        50, 0.2F, 0.3F)
+        };
+        spell.release.particles_scaled_with_ranged = new ParticleBatch[]{
+                new ParticleBatch(SpellEngineParticles.area_effect_658.id().toString(),
+                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.GROUND,
+                        1, 0, 0)
+                        .scale(0.5F)
+                        .color(ORANGE.toRGBA())
+        };
+
+        spell.target.type = Spell.Target.Type.AREA;
+        spell.target.area = new Spell.Target.Area();
+        spell.target.area.angle_degrees = 360;
+
+        var damage = SpellBuilder.Impacts.damage(0.5F, 0F);
+        damage.particles = new ParticleBatch[]{
+                new ParticleBatch(SpellEngineParticles.smoke_medium.id().toString(),
+                        ParticleBatch.Shape.PIPE, ParticleBatch.Origin.CENTER,
+                        20, 0.1F, 3.0F)
+        };
+        damage.sound = new Sound("forcemaster_rpg:knuckle_attack");
+
+        var damageArcane = SpellBuilder.Impacts.damage(0.2F, 0F);
+        damageArcane.school = SpellSchools.ARCANE;
 
         var custom = new Spell.Impact();
-
+        bossDeny(custom);
         custom.action = new Spell.Impact.Action();
         custom.action.custom = new Spell.Impact.Action.Custom();
-        bossDeny(custom);
         custom.action.type = Spell.Impact.Action.Type.CUSTOM;
         custom.action.custom.intent = SpellTarget.Intent.HARMFUL;
-        custom.action.custom.handler = "more_rpg_classes:knock_up_fixed";
-        custom.particles = new ParticleBatch[]{
-                new ParticleBatch(
-                        SpellEngineParticles.smoke_medium.id().toString(),
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.FEET,
-                        20, 0.1F, 0.3F)
-                        .extent(0.25F)
-                        .color(Color.WHITE.toRGBA()),
-        };
-        spell.impacts = List.of(custom);
+        custom.action.custom.handler = "more_rpg_classes:knock_up";
 
-        SpellBuilder.Cost.cooldown(spell, 4F);
+        spell.impacts = List.of(damage, damageArcane, custom);
+        SpellBuilder.Cost.cooldown(spell, 15);
+        spell.cost.cooldown.attempt_duration = 1.5F;
 
         return new Entry(id, spell, title, description, null);
     }
+    ///PASSIVES
     public static final Entry knuckle_arcane_overflow = add(knuckle_arcane_overflow());
     private static Entry knuckle_arcane_overflow() {
         var id = Identifier.of(MOD_ID, "knuckle_arcane_overflow");
@@ -154,9 +186,9 @@ public class ForcemasterSpells {
         var spell = SpellBuilder.createSpellPassive();
         spell.tooltip = new Spell.Tooltip();
         spell.tooltip.show_header = false;
-        spell.tooltip.name = new Spell.Tooltip.LineOptions(false, false);
+        spell.tooltip.name = new Spell.Tooltip.LineOptions(true, false);
         spell.tooltip.description.color = Formatting.DARK_GREEN.asString();
-        spell.tooltip.description.show_in_compact = false;
+        spell.tooltip.description.show_in_compact = true;
         spell.school = SpellSchools.ARCANE;
         spell.range = 0;
 
@@ -247,10 +279,10 @@ public class ForcemasterSpells {
         var description = "{trigger_chance_1} chance for {stash_duration} seconds to stun targets on damage, only with fist weapons.";
         var spell = activeSpellBase();
         spell.school = SpellSchools.ARCANE;
-        spell.tier = 1;
+        spell.tier = 2;
         spell.range = 2;
 
-        spell.release.animation = "forcemaster_rpg:stonehand_cast";
+        spell.release.animation = PlayerAnimation.of("forcemaster_rpg:stonehand_cast");
         spell.release.sound = Sound.withVolume(Identifier.of("forcemaster_rpg:stonehand_cast"), 0.35F);
         spell.release.particles = new ParticleBatch[]{
                 new ParticleBatch("more_rpg_classes:stone_particle",
@@ -300,62 +332,59 @@ public class ForcemasterSpells {
 
         spell.impacts = List.of(custom);
         configureCooldown(spell, 20, 0.3F);
-        return new Entry(id, spell, title, description, null);
+        return new Entry(id, spell, title, description, null).book(Book.FORCEMASTER);
     }
     public static Entry belial_smashing = add(belial_smashing());
     private static Entry belial_smashing() {
         var id = Identifier.of(MOD_ID, "belial_smashing");
-        var effect = ForcemasterEffects.ARCANE_OVERFLOW;
         var title = "Belial Smashing";
-        var description = "Charges to the targets direction, and dealing arcane {damage_1} & {damage_2} physical-damage.";
-        var spell = activeSpellBase();
+        var description = "Charges to the targets direction, punching all enemies in your path.";
+        var spell = SpellBuilder.createSpellActive();
         spell.school = SpellSchools.ARCANE;
         spell.tier = 3;
-        spell.range = 8.5F;
 
-        spell.target.type = Spell.Target.Type.AIM;
-        spell.target.aim = new Spell.Target.Aim();
-        spell.target.aim.sticky = true;
+        SpellBuilder.Casting.instant(spell);
+        SpellBuilder.Target.none(spell);
 
-        spell.release.animation = "forcemaster_rpg:fist_rush";
-        spell.release.sound = Sound.withVolume(Identifier.of("entity.generic.explode"), 0.35F);
         spell.release.particles = new ParticleBatch[]{
                 new ParticleBatch(SpellEngineParticles.smoke_large.id().toString(),
                         ParticleBatch.Shape.WIDE_PIPE, ParticleBatch.Origin.FEET,
                         35, 0.1F, 0.5F).followEntity(true).color(Color.WHITE.toRGBA())
         };
 
-        var damage = SpellBuilder.Impacts.damage(0.75F,0.5F);
-        damage.particles = new ParticleBatch[]{
-                new ParticleBatch(SpellEngineParticles.smoke_large.id().toString(),
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.CENTER,
-                        15, 0.1F, 0.3F).color(Color.WHITE.toRGBA())
-        };
-        damage.sound = new Sound("forcemaster_rpg:knuckle_attack");
+        var attack = new Spell.Delivery.Melee.Attack();
+        attack.attack_speed_multiplier = 1.2F;
+        attack.delay = 0.1F;
+        attack.hitbox = new Spell.Delivery.Melee.HitBox();
+        attack.hitbox.arc = 160;
+        attack.hitbox.height = 0.5F;
+        attack.hitbox.width = 0.5F;
+        attack.forward_momentum = 3.0F;
+        attack.movement_slipperiness = 0.2F;
 
-        var damagePhysical = SpellBuilder.Impacts.damage(0.5F,0F);
-        damagePhysical.school = ExternalSpellSchools.PHYSICAL_MELEE;
+        attack.additional_strikes = 8;
+        attack.additional_strike_delay = 0.15F;
+        attack.additional_hits_on_same_target = false;
+        attack.animation = PlayerAnimation.of("forcemaster_rpg:fist_rush");
+        attack.animation.speed = 1F;
+        attack.swing_sound = Sound.of(SpellEngineSounds.WEAPON_HAMMER_SWING.id());
+        attack.impact_sound = Sound.of(ModSounds.KNUCKLE_ATTACK.id());
 
-        var buff = SpellBuilder.Impacts.effectAdd(effect.id.toString(), 10,1,5);
-        buff.action.status_effect.refresh_duration = true;
+        var damageArcane = SpellBuilder.Impacts.damage(0.3F,0F);
+        damageArcane.school  = SpellSchools.ARCANE;
 
-        var custom = new Spell.Impact();
-        custom.action = new Spell.Impact.Action();
-        custom.action.custom = new Spell.Impact.Action.Custom();
-        custom.action.type = Spell.Impact.Action.Type.CUSTOM;
-        custom.action.custom.intent = SpellTarget.Intent.HARMFUL;
-        custom.action.custom.handler = "more_rpg_classes:rush_forward_to_target";
+        spell.impacts = List.of(damageArcane);
+        SpellBuilder.Deliver.melee(spell, List.of(attack));
+        spell.deliver.melee.allow_airborne = false;
 
-        spell.impacts = List.of(custom, damage,damagePhysical, buff);
-        configureCooldown(spell, 22, 0.5F);
-        return new Entry(id, spell, title, description, null);
+        SpellBuilder.Cost.cooldown(spell, 23);
+        return new Entry(id, spell, title, description, null).book(Book.FORCEMASTER);
     }
-
     public static final Entry asal = add(asal());
     private static Entry asal() {
         var id = Identifier.of(MOD_ID, "asal");
         var title = "Asalraalaikum";
-        var description = "A devastating charged strike that deals massive {damage_1} arcane & {damage_2} physical-damage. Consumes Arcane Overflow and exhausts the caster.";
+        var description = "A devastating charged punch that deals massive {damage} damage. Consumes Arcane Overflow and exhausts the caster.";
         var spell = activeSpellBase();
         spell.school = SpellSchools.ARCANE;
         spell.tier = 4;
@@ -363,7 +392,7 @@ public class ForcemasterSpells {
 
         spell.active.cast.movement_speed = 0.1F;
         spell.active.cast.duration = 0.75F;
-        spell.active.cast.animation = "forcemaster_rpg:asal_cast";
+        spell.active.cast.animation = PlayerAnimation.of("forcemaster_rpg:asal_cast");
         spell.active.cast.sound = new Sound("spell_engine:generic_arcane_casting");
         spell.active.cast.particles = new ParticleBatch[]{
                 new ParticleBatch(SpellEngineParticles.MagicParticles.get(
@@ -373,11 +402,6 @@ public class ForcemasterSpells {
                         10, 0.2F, 0.6F).extent(0.5F).color(Color.ARCANE.toRGBA())
         };
 
-        spell.target.type = Spell.Target.Type.AREA;
-        spell.target.area = new Spell.Target.Area();
-        spell.target.area.angle_degrees = 50;
-
-        spell.release.animation = "forcemaster_rpg:asal_release";
         spell.release.sound = Sound.withVolume(Identifier.of("forcemaster_rpg:asal_release"), 0.35F);
         spell.release.particles = new ParticleBatch[]{
                 new ParticleBatch(SpellEngineParticles.smoke_medium.id().toString(),
@@ -385,6 +409,19 @@ public class ForcemasterSpells {
                         ParticleBatch.Rotation.LOOK,
                         15, 1.0F, 10.0F, 360, 2).color(Color.ARCANE.toRGBA())
         };
+
+        SpellBuilder.Target.none(spell);
+
+        var punch = new Spell.Delivery.Melee.Attack();
+        punch.attack_speed_multiplier = 2.0F;
+        punch.delay = 0.25F;
+        punch.hitbox = new Spell.Delivery.Melee.HitBox();
+        punch.hitbox.height = 2.0F;
+        punch.hitbox.length = 6.5F;
+        punch.hitbox.width = 1.5F;
+        punch.animation = PlayerAnimation.of("forcemaster_rpg:asal_release");
+
+        SpellBuilder.Deliver.melee(spell, List.of(punch));
 
         var damage = SpellBuilder.Impacts.damage(3.0F, 2.0F);
         damage.particles = new ParticleBatch[]{
@@ -394,10 +431,7 @@ public class ForcemasterSpells {
         };
         damage.sound = Sound.withVolume(Identifier.of("entity.dragon_fireball.explode"), 0.5F);
 
-        var damagePhysical = SpellBuilder.Impacts.damage(1.5F, 0F);
-        damagePhysical.school = ExternalSpellSchools.PHYSICAL_MELEE;
-
-        spell.impacts = List.of(damage, damagePhysical);
+        spell.impacts = List.of(damage);
 
         spell.cost.effect_id = ForcemasterEffects.ARCANE_OVERFLOW.id.toString();
         spell.cost.exhaust = 40F;
@@ -406,77 +440,14 @@ public class ForcemasterSpells {
         spell.cost.cooldown.duration = 30;
         spell.cost.cooldown.haste_affected = true;
 
-        return new Entry(id, spell, title, description, null);
+        return new Entry(id, spell, title, description, null).book(Book.FORCEMASTER);
     }
-
-    public static final Entry burstcrack = add(burstcrack());
-    private static Entry burstcrack() {
-        var id = Identifier.of(MOD_ID, "burstcrack");
-        var title = "Burstcrack";
-        var description = "Unleashes a shockwave around you, knocking up enemies and dealing {damage} arcane and physical damage.";
-        var spell = activeSpellBase();
-        spell.school = SpellSchools.ARCANE;
-        spell.tier = 2;
-        spell.range = 5.5F;
-
-        spell.active.cast.movement_speed = 0.5F;
-        spell.active.cast.duration = 1.0F;
-        spell.active.cast.animation = "forcemaster_rpg:burstcrack_cast";
-
-        spell.target.type = Spell.Target.Type.AREA;
-        spell.target.area = new Spell.Target.Area();
-        spell.target.area.angle_degrees = 360;
-
-        spell.release.animation = "forcemaster_rpg:burstcrack_release";
-        spell.release.sound = Sound.withVolume(Identifier.of("entity.generic.explode"), 0.4F);
-        spell.release.particles = new ParticleBatch[]{
-                new ParticleBatch(SpellEngineParticles.smoke_medium.id().toString(),
-                        ParticleBatch.Shape.CIRCLE, ParticleBatch.Origin.CENTER,
-                        50, 0.2F, 0.3F)
-        };
-        spell.release.particles_scaled_with_ranged = new ParticleBatch[]{
-                new ParticleBatch(SpellEngineParticles.area_effect_658.id().toString(),
-                        ParticleBatch.Shape.SPHERE, ParticleBatch.Origin.GROUND,
-                        1, 0, 0)
-                        .scale(0.5F)
-                        .color(ORANGE.toRGBA())
-        };
-
-        var custom = new Spell.Impact();
-        bossDeny(custom);
-        custom.action = new Spell.Impact.Action();
-        custom.action.custom = new Spell.Impact.Action.Custom();
-        custom.action.type = Spell.Impact.Action.Type.CUSTOM;
-        custom.action.custom.intent = SpellTarget.Intent.HARMFUL;
-        custom.action.custom.handler = "more_rpg_classes:knock_up";
-
-        var damage = SpellBuilder.Impacts.damage(0.5F, 0F);
-        damage.particles = new ParticleBatch[]{
-                new ParticleBatch(SpellEngineParticles.smoke_medium.id().toString(),
-                        ParticleBatch.Shape.PIPE, ParticleBatch.Origin.CENTER,
-                        20, 0.1F, 3.0F)
-        };
-        damage.sound = new Sound("forcemaster_rpg:knuckle_attack");
-
-        var damagePhysical = SpellBuilder.Impacts.damage(0.5F, 0F);
-        damagePhysical.school = ExternalSpellSchools.PHYSICAL_MELEE;
-
-        spell.impacts = List.of(custom, damage, damagePhysical);
-
-        spell.cost.exhaust = 0.3F;
-        spell.cost.durability = 1;
-        spell.cost.cooldown = new Spell.Cost.Cooldown();
-        spell.cost.cooldown.duration = 15;
-        spell.cost.cooldown.haste_affected = false;
-
-        return new Entry(id, spell, title, description, null);
-    }
-
-    public static final Entry improved_burstcrack = add(improved_burstcrack());
-    private static Entry improved_burstcrack() {
-        var id = Identifier.of(MOD_ID, "improved_burstcrack");
-        var title = "Improved Burstcrack";
-        var description = "Increases the range of Burstcrack by {range_add}";
+    /// MODIFIER
+    public static final Entry improved_belial_smashing = add(improved_belial_smashing());
+    private static Entry improved_belial_smashing() {
+        var id = Identifier.of(MOD_ID, "improved_belial_smashing");
+        var title = "Improved Belial Smashing";
+        var description = "Reduces the cooldown of Belial Smashing by {cooldown_duration_deduct} sec.";
         var spell = new Spell();
         spell.school = SpellSchools.ARCANE;
         spell.range = 0;
@@ -491,8 +462,8 @@ public class ForcemasterSpells {
         spell.tooltip.description.show_in_compact = true;
 
         var modifier = new Spell.Modifier();
-        modifier.spell_pattern = "forcemaster_rpg:burstcrack";
-        modifier.range_add = 2.0F;
+        modifier.spell_pattern = "forcemaster_rpg:belial_smashing";
+        modifier.cooldown_duration_deduct = 3;
         spell.modifiers = List.of(modifier);
 
         return new Entry(id, spell, title, description, null);